+++ packages/eslint-config-cli/example.js
@@ -1,9 +1,8 @@
 /* require should be allowed */
-const fs = require("fs");
+const fs = require("node:fs");
 
 /* Promise and ES6+ should be allowed */
-export const myFunc = () =>
-    Promise.resolve(fs.readFileSync("foo.txt", "utf-8"));
+export const myFunc = () => Promise.resolve(fs.readFileSync("foo.txt", "utf8"));
 
 /* console should be allowed */
 console.log(myFunc());