+++ packages/eslint-config-jest/index.mjs
@@ -30,9 +30,9 @@ function merge(result, it) {
 const recommended = jestPlugin.configs["flat/recommended"];
 const style = jestPlugin.configs["flat/style"];
 
-const config = defineConfig({
+const defaultConfig = defineConfig({
     name: "@forsakringskassan/eslint-config-jest",
-    files: ["**/*.spec.[jt]s"],
+    files: ["**/*.spec.{js,ts,cjs,mjs,mts}"],
     ignores: ["cypress/**", "tests/e2e/**"],
 
     languageOptions: {
@@ -69,6 +69,19 @@ const config = defineConfig({
 
         /* jest uses @jest-* tags for per-file configuration */
         "tsdoc/syntax": "off",
+
+        "unicorn/consistent-function-scoping": "off",
+
+        /* allow either kebab-case or PascalCase for testcases, it should match the file under test */
+        "unicorn/filename-case": [
+            "error",
+            {
+                cases: {
+                    kebabCase: true,
+                    pascalCase: true,
+                },
+            },
+        ],
     },
 });
 
@@ -76,4 +89,5 @@ const config = defineConfig({
  * @param {Config} [override]
  * @returns {Config}
  */
-export default (override) => merge(config, override ?? {});
+const config = (override) => merge(defaultConfig, override ?? {});
+export default config;