+++ packages/eslint-config-vue/ExampleComponent.pageobject.ts
@@ -1,0 +1,2 @@
+/* pageobjects should have the PascalCase filename case */
+export {};