+++ packages/eslint-config-vue/ExampleComponent.selectors.ts
@@ -1,0 +1,2 @@
+/* selector objects should have the PascalCase filename case */
+export {};