+++ packages/eslint-config/eslint.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
-const { spawn } = require("child_process");
-const path = require("path");
+const { spawn } = require("node:child_process");
+const path = require("node:path");
 
 const pkgPath = path.dirname(require.resolve("eslint/package.json"));
 const binary = path.join(pkgPath, "bin/eslint");